requirements.txt
tensorflow-cpu==2.3.0
Flask==1.1.2
Werkzeug==1.0.1
